fix: 통합테스트 추가 및 단위테스트 수정

+++ src/main/kotlin/com/mss/assignment/domain/brand/BrandService.kt
@@ -1,5 +1,6 @@
 package com.mss.assignment.domain.brand
 
+import com.mss.assignment.domain.product.ProductRepository
 import com.mss.assignment.dto.BrandDto
 import com.mss.assignment.exception.ErrorCode
 import com.mss.assignment.exception.GlobalHttpException
@@ -12,11 +13,12 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 class BrandService(
     private val brandRepository: BrandRepository,
+    private val productRepository: ProductRepository,
 ) {
     @Transactional
     fun createBrand(name: String): BrandDto {
         if (brandRepository.existsByName(name)) {
-            throw GlobalHttpException(HttpStatus.CONFLICT, ErrorCode.CONFLICT_NAME)
+            throw GlobalHttpException(HttpStatus.CONFLICT, ErrorCode.NAME_CONFLICT)
         }
         val brand = Brand(name = name)
         return BrandDto.fromEntity(brandRepository.save(brand))
@@ -29,7 +31,7 @@ class BrandService(
     ): BrandDto {
         val brand = getBrandById(id)
         if (brandRepository.existsByName(name)) {
-            throw GlobalHttpException(HttpStatus.CONFLICT, ErrorCode.CONFLICT_NAME)
+            throw GlobalHttpException(HttpStatus.CONFLICT, ErrorCode.NAME_CONFLICT)
         }
         brand.updateName(name)
         return BrandDto.fromEntity(brandRepository.save(brand))
@@ -46,9 +48,15 @@ class BrandService(
         allEntries = true,
     )
     fun deleteBrand(id: Long) {
+        // 브랜드 존재 여부 확인
         if (!brandRepository.existsById(id)) {
             throw NotFoundException(ErrorCode.BRAND_NOT_FOUND)
         }
+        // 해당 브랜드가 상품을 가지고 있는지 확인
+        if (productRepository.existsByBrandId(id)) {
+            throw GlobalHttpException(HttpStatus.CONFLICT, ErrorCode.BRAND_HAS_PRODUCTS)
+        }
+
         brandRepository.deleteById(id)
     }
 }
